Practica 1-A: try to reduce the version of openGL to use it in WSL

+++ pr1a/Source/Rendering/DrawPolygon.cpp
@@ -15,7 +15,7 @@ AlgGeom::DrawPolygon::DrawPolygon(Polygon& polygon)
         Point point = _polygon.getVertexAt(vertexIdx).getPoint();
 
         component->_vertices.push_back(VAO::Vertex {vec3(point.getX(), point.getY(), .0f)});
-        component->_indices[VAO::IBO_LINE].insert(component->_indices[VAO::IBO_LINE].end(), {vertexIdx, (vertexIdx + 1) % numVertices, RESTART_PRIMITIVE_INDEX});
+        component->_indices[VAO::IBO_LINE].insert(component->_indices[VAO::IBO_LINE].end(), {vertexIdx, static_cast<unsigned int>((vertexIdx + 1) % numVertices), RESTART_PRIMITIVE_INDEX});
     }
 
     this->_components.push_back(std::unique_ptr<Component>(component));
